+++ lib/src/db/lib_prod_archives.dart
@@ -6,6 +6,7 @@ import 'package:elec_server/src/db/cme/cme_energy_settlements.dart';
 import 'package:elec_server/src/db/config.dart';
 import 'package:elec_server/src/db/ieso/rt_generation.dart';
 import 'package:elec_server/src/db/ieso/rt_zonal_demand.dart';
+import 'package:elec_server/src/db/isoexpress/da_energy_offer.dart';
 import 'package:elec_server/src/db/polygraph/polygraph_archive.dart';
 import 'package:elec_server/src/db/utilities/eversource/supplier_backlog_rates.dart';
 import 'package:elec_server/src/db/utilities/retail_suppliers_offers_archive.dart';
@@ -20,6 +21,20 @@ CmeSettlementsEnergyArchive getCmeEnergySettlementsArchive() {
   }
   return CmeSettlementsEnergyArchive(dbConfig: dbConfig, dir: dir);
 }
+
+DaEnergyOfferArchive getDaEnergyOfferArchive() {
+  var dbConfig = ComponentConfig(
+      host: '127.0.0.1', dbName: 'isoexpress', collectionName: 'da_energy_offer');
+  var dir =
+      '${Platform.environment['HOME'] ?? ''}/Downloads/Archive/Downloads/'
+      'Archive/IsoExpress/DaEnergyOffer/Raw/';
+  if (!Directory(dir).existsSync()) {
+    Directory(dir).createSync(recursive: true);
+  }
+  return DaEnergyOfferArchive(dbConfig: dbConfig, dir: dir);
+
+}
+
 
 IesoRtGenerationArchive getIesoRtGenerationArchive() {
   var dbConfig = ComponentConfig(